extract our save as example flow
